Added dice removal logic
- Extracted `useBoard` out of `Board`
- Exposed `removeDiceFromColumn` method
- Called method when a dice is placed in `App`
- Updated `yarn.lock`
