+++ src/components/App.tsx
@@ -1,11 +1,12 @@
 import * as React from 'react'
+import { useParams } from 'react-router-dom'
+import { useBoard } from '../hooks/useBoard'
 import { useTurn } from '../hooks/useTurn'
 import { useNames } from '../hooks/useNames'
 import { useGame, useResumeGame } from '../hooks/useGame'
 import { getRandomDice } from '../utils/random'
 import { connectToAbly } from '../utils/connectToAbly'
-import { Board, useBoard } from './Board'
-import { useParams } from 'react-router-dom'
+import { Board } from './Board'
 
 connectToAbly()
 
@@ -20,11 +21,19 @@ export function App() {
   const roomId = `${roomName}:${roomKey}`
   const [dice, setDice] = React.useState(getRandomDice())
 
-  const { columns: opponentColumns, addToColumn: addToOpponentColumn } =
-    useBoard()
-  const { columns, addToColumn } = useBoard({
+  const {
+    columns: opponentColumns,
+    addToColumn: addToOpponentColumn,
+    removeDiceFromColumn: removeDiceFromOpponentColumn
+  } = useBoard()
+  const {
+    columns: myColumns,
+    addToColumn: addToMyColumn,
+    removeDiceFromColumn: removeDiceFromMyColumn
+  } = useBoard({
     onDicePlaced(column, value) {
       sendPlay({ column, value })
+      removeDiceFromOpponentColumn(column, value)
       setDice(getRandomDice())
     }
   })
@@ -32,16 +41,19 @@ export function App() {
   const { sendPlay } = useGame(roomId, {
     onOpponentPlay({ column, value }) {
       addToOpponentColumn(column, value)
+      removeDiceFromMyColumn(column, value)
     }
   })
   const myTurn = useTurn(roomId)
   const { myName, opponentName } = useNames(roomId)
   useResumeGame(roomId, {
     onMyPlay({ column, value }) {
-      addToColumn(column, value, false)
+      addToMyColumn(column, value, false)
+      removeDiceFromOpponentColumn(column, value)
     },
     onOpponentPlay({ column, value }) {
       addToOpponentColumn(column, value, false)
+      removeDiceFromMyColumn(column, value)
     }
   })
 
@@ -58,8 +70,8 @@ export function App() {
       </div>
       <Board isOpponentBoard columns={opponentColumns} name={opponentName} />
       <Board
-        columns={columns}
-        onColumnClick={(colIndex) => addToColumn(colIndex, dice)}
+        columns={myColumns}
+        onColumnClick={(colIndex) => addToMyColumn(colIndex, dice)}
         nextDie={dice}
         canPlay={myTurn}
         name={myName}
